feat(integration-tests): seperate scenarios
