+++ integration-tests/Cargo.toml
@@ -8,8 +8,16 @@ homepage.workspace = true
 repository.workspace = true
 
 [[bin]]
-name = "integration-tests"
-path = "bin/scenario.rs"
+name = "witness_scenario"
+path = "bin/witness_scenario.rs"
+
+[[bin]]
+name = "proof_scenario"
+path = "bin/proof_scenario.rs"
+
+# [[bin]]
+# name = "unit_tests"
+# path = "bin/unit_test.rs"
 
 [dependencies]
 anyhow.workspace = true
