+++ integration-tests/bin/witness_scenario.rs
@@ -1,9 +1,8 @@
 use alloy_primitives::{b256, B256};
 use anyhow::Result;
 use clap::Parser;
-use integration_tests::{ProverRequest, TestClient, WitnessRequest};
+use integration_tests::{TestClient, WitnessRequest};
 use kroma_witnessgen::errors::ErrorCode as WitnessErrorCode;
-use serde::{Deserialize, Serialize};
 use std::{fs::File, io::Write, thread::sleep, time::Duration};
 
 #[derive(Parser, Debug)]
@@ -14,13 +13,9 @@ struct Args {
 
     #[clap(short, long)]
     l1_head_hash: B256,
-}
 
-#[derive(Debug, Serialize, Deserialize)]
-struct ProofFixture {
-    program_key: String,
-    public_values: String,
-    proof: String,
+    #[clap(short, long)]
+    witness_data: String,
 }
 
 #[tokio::main]
@@ -29,11 +24,6 @@ async fn main() -> Result<()> {
     let client = TestClient::new();
 
     let _ = client.witnessgen_spec().await;
-    let _ = client.prover_spec().await;
-
-    ////////////////////////////////////////////////////////////////
-    //                   WITNESS GENERATOR TEST                   //
-    ////////////////////////////////////////////////////////////////
 
     // The response should be `Processing`.
     let request_result = client.request_witness(args.l2_hash, args.l1_head_hash).await.unwrap();
@@ -63,40 +53,10 @@ async fn main() -> Result<()> {
         TestClient::execute_witness(&witness_result).await.expect("Failed to execute witness");
     tracing::info!("Witness execution succeeded: {:?}", report.total_instruction_count());
 
-    ////////////////////////////////////////////////////////////////
-    //                        PROVER TEST                         //
-    ////////////////////////////////////////////////////////////////
-
-    // The response should be `Processing`.
-    let request_result =
-        client.request_prove(args.l2_hash, args.l1_head_hash, &witness_result).await.unwrap();
-    assert_eq!(request_result, ProverRequest::Processing);
-
-    // The same response is returned for the same request.
-    let request_result =
-        client.request_prove(args.l2_hash, args.l1_head_hash, &witness_result).await.unwrap();
-    assert_eq!(request_result, ProverRequest::Processing);
-
-    let proof_result = loop {
-        let proof_result = client.get_proof(args.l2_hash, args.l1_head_hash).await;
-        if proof_result.request_status == ProverRequest::Completed {
-            break proof_result;
-        }
-        if let ProverRequest::Failed(_) = proof_result.request_status {
-            panic!("Failed to get witness");
-        }
-        sleep(Duration::from_secs(20));
-    };
-
-    let proof_fixture = ProofFixture {
-        program_key: proof_result.program_key,
-        public_values: proof_result.public_values,
-        proof: proof_result.proof,
-    };
-    let proof_json = serde_json::to_string_pretty(&proof_fixture)?;
-    let mut file = File::create("output.json")?;
-    file.write_all(proof_json.as_bytes())?;
-    tracing::info!("Proof was saved");
+    let witness_json = serde_json::to_string_pretty(&witness_result)?;
+    let mut file = File::create(args.witness_data)?;
+    file.write_all(witness_json.as_bytes())?;
+    tracing::info!("Witness was saved");
 
     Ok(())
 }
